requirements.txt
pandas>=1.5.0
numpy>=1.23.0
matplotlib>=3.6.0
seaborn>=0.12.0
openpyxl>=3.0.0
faker>=15.0.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
